Class içindeki decimal değişkenler double olarak güncellendi

+++ AHAFit-Data/Food.cs
@@ -10,10 +10,10 @@ namespace AHAFit_Data
     {
         public int FoodId { get; set; }
         public string Name { get; set; }
-        public decimal Calorie { get; set; }
-        public decimal Carbohydrate { get; set; }
-        public decimal Protein { get; set; }
-        public decimal Fat { get; set; }
+        public double Calorie { get; set; }
+        public double Carbohydrate { get; set; }
+        public double Protein { get; set; }
+        public double Fat { get; set; }
         public string PhotoURL { get; set; }
         public List<MemberFood> MemberFoods { get; set; }
         public List<MealFood> MealFoods { get; set; }
